unclutter init pkg
